+++ scalyr_agent/monitor_utils/k8s.py
@@ -2,6 +2,9 @@
 import hashlib
 import logging
 import os
+import random
+import string
+
 import scalyr_agent.monitor_utils.annotation_config as annotation_config
 import scalyr_agent.third_party.requests as requests
 import scalyr_agent.json_lib as json_lib
@@ -17,6 +20,12 @@ import urllib
 global_log = scalyr_logging.getLogger(__name__)
 
 class K8sApiException( Exception ):
+    """A wrapper around Exception that makes it easier to catch k8s specific
+    exceptions
+    """
+    pass
+
+class KubeletApiException( Exception ):
     """A wrapper around Exception that makes it easier to catch k8s specific
     exceptions
     """
@@ -768,11 +777,40 @@ class KubernetesApi( object ):
         metadata = pod.get( 'metadata', {} )
         annotations = metadata.get( 'annotations', {} )
 
-        result = None
         if 'agent.config.scalyr.com/cluster_name' in annotations:
-            result = annotations['agent.config.scalyr.com/cluster_name']
+            return annotations['agent.config.scalyr.com/cluster_name']
 
-        return result
+        # If the user did not specify any cluster name, we need to supply a default that will be the same for all
+        # other scalyr agents connected to the same cluster.  Unfortunately, k8s does not actually supply the cluster
+        # name via any API, so we must make one up.
+        # We create a random string using the creation timestamp of the default timestamp as a seed.  The idea is that
+        # that creation timestamp should never change and all agents connected to the cluster will see the same value
+        # for that seed.
+        namespaces = self.query_namespaces()
+
+        # Get the creation timestamp from the default namespace.  We try to be very defensive in case the API changes.
+        if namespaces and 'items' in namespaces:
+            for item in namespaces['items']:
+                if 'metadata' in item and 'name' in item['metadata'] and item['metadata']['name'] == 'default':
+                    if 'creationTimestamp' in item['metadata']:
+                        return 'k8s-cluster-%s' % self.__create_random_string(item['metadata']['creationTimestamp'], 6)
+        return None
+
+    def __create_random_string(self, seed_string, num_chars):
+        """
+        Return a random string of num_char characters, composed of uppercase characters and digits.
+
+        @param seed_string: The seed to use when creating the psrng
+        @param num_chars: The desired size of the string.
+
+        @type seed_string: str
+        @type num_chars: int
+        @return: A random string
+        @rtype: str
+        """
+        prng = random.Random(abs(hash(seed_string)))
+        return ''.join(prng.choice(string.ascii_lowercase + string.digits) for _ in range(num_chars))
+
 
 
     def query_api( self, path, pretty=0 ):
@@ -855,4 +893,46 @@ class KubernetesApi( object ):
     def query_namespaces( self ):
         """Wrapper to query all namespaces"""
         return self.query_api( '/api/v1/namespaces' )
+
+class KubeletApi( object ):
+    """
+        A class for querying the kubelet API
+    """
+
+    def __init__( self, k8s, port=10255 ):
+        """
+        @param k8s - a KubernetesApi object
+        """
+        pod_name = k8s.get_pod_name()
+        pod = k8s.query_pod( k8s.namespace, pod_name )
+        spec = pod.get( 'spec', {} )
+        status = pod.get( 'status', {} )
+
+        host_ip = status.get( 'hostIP', None )
+
+        if host_ip is None:
+            raise KubeletApiException( "Unable to get host IP for pod: %s/%s" % (k8s.namespace, pod_name) )
+
+        self._session = requests.Session()
+        headers = {
+            'Accept': 'application/json',
+        }
+        self._session.headers.update( headers )
+
+        self._http_host = "http://%s:%d" % ( host_ip, port )
+        self._timeout = 10.0
+
+    def query_api( self, path ):
+        """ Queries the kubelet API at 'path', and converts OK responses to JSON objects
+        """
+        url = self._http_host + path
+        response = self._session.get( url, timeout=self._timeout )
+        if response.status_code != 200:
+            global_log.log(scalyr_logging.DEBUG_LEVEL_3, "Invalid response from Kubelet API.\n\turl: %s\n\tstatus: %d\n\tresponse length: %d"
+                % ( url, response.status_code, len(response.text)), limit_once_per_x_secs=300, limit_key='kubelet_api_query' )
+            raise KubeletApiException( "Invalid response from Kubelet API when querying '%s': %s" %( path, str( response ) ) )
+        return json_lib.parse( response.text )
+
+    def query_stats( self ):
+        return self.query_api( '/stats/summary' )