+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 Scalyr Agent 2 Changes By Release
 =================================
 
+## 2.0.37 "Ly-Cilph" - Oct 19, 2018
+
+<!---
+Packaged by Steven Czerwinski <czerwin@scalyr.com> on Sep 19, 2018 15:00 -0700
+--->
+
+Features
+
+* Finalized new support for Kubernetes.  Not turned on by default yet.  Will be generally available in November.
+
+Bugs
+* Fix journald monitor to system default path.
+
 ## 2.0.36 "Kliint" - Sept 21, 2018
 
 <!---